use consumer instead of property
git-svn-id: svn://svn.code.sf.net/p/jsidplay2/code/trunk@1178 05cac4c1-a8e4-46c3-9d4b-80028926d4ef

+++ jsidplay2/src/main/java/ui/directory/Directory.java
@@ -95,10 +95,14 @@ public class Directory extends AnchorPane implements UIPart {
 		try {
 			DirectoryItem dirItem = directory.getSelectionModel()
 					.getSelectedItem();
+			if (dirItem == null) {
+				return;
+			}
 			DirEntry dirEntry = dirItem.getDirEntry();
 			if (dirEntry != null) {
-				File autoStartFile = new File(util.getConfig().getSidplay2Section()
-						.getTmpDir(), dirEntry.getValidFilename() + ".prg");
+				File autoStartFile = new File(util.getConfig()
+						.getSidplay2Section().getTmpDir(),
+						dirEntry.getValidFilename() + ".prg");
 				autoStartFile.deleteOnExit();
 				dirEntry.save(autoStartFile);
 				autoStartFileProperty.set(autoStartFile);
